refactor!: remove exporting internal APIs (#14583)
