+++ packages/vite/src/node/index.ts
@@ -1,7 +1,12 @@
 import type * as Rollup from 'rollup'
 
 export type { Rollup }
-export * from './config'
+export {
+  defineConfig,
+  loadConfigFromFile,
+  resolveConfig,
+  sortUserPlugins,
+} from './config'
 export { createServer } from './server'
 export { preview } from './preview'
 export { build } from './build'
@@ -12,6 +17,23 @@ export { buildErrorMessage } from './server/middlewares/error'
 export * from './publicUtils'
 
 // additional types
+export type {
+  AppType,
+  ConfigEnv,
+  ExperimentalOptions,
+  InlineConfig,
+  LegacyOptions,
+  PluginHookUtils,
+  PluginOption,
+  ResolveFn,
+  ResolvedWorkerOptions,
+  ResolvedConfig,
+  UserConfig,
+  UserConfigExport,
+  UserConfigFn,
+  UserConfigFnObject,
+  UserConfigFnPromise,
+} from './config'
 export type { FilterPattern } from './utils'
 export type { CorsOptions, CorsOrigin, CommonServerOptions } from './http'
 export type {
@@ -28,6 +50,9 @@ export type {
   LibraryFormats,
   RenderBuiltAssetUrl,
   ResolvedBuildOptions,
+  ModulePreloadOptions,
+  ResolvedModulePreloadOptions,
+  ResolveModulePreloadDependenciesFn,
 } from './build'
 export type {
   PreviewOptions,
@@ -39,10 +64,7 @@ export type {
   DepOptimizationMetadata,
   DepOptimizationOptions,
   DepOptimizationConfig,
-  DepOptimizationResult,
-  DepOptimizationProcessing,
   OptimizedDepInfo,
-  DepsOptimizer,
   ExportsData,
 } from './optimizer'
 export type {
